+++ Example_Project/Scaler_node.py
@@ -1,6 +1,6 @@
 from PySide6 import QtWidgets
 
-from node_editor.gui.node import Node
+from node_editor.node import Node
 from Example_Project.common_widgets import FloatLineEdit
